requirements.txt
streamlit
pandas
selenium
requests
beautifulsoup4
google-genai
fastapi
uvicorn
pydantic
python-multipart
python-dotenv